requirements.txt
fastapi
uvicorn
playwright
python-dotenv
supabase
httpx
